requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
drf-nested-routers==0.94.2
Pillow==11.3.0 